tests/wpt/tests/navigation-api/navigate-event/navigate-destination-dynamic-index.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../navigation-methods/return-value/resources/helpers.js"></script>
<script>
promise_test(async t => {
  let start_index = navigation.currentEntry.index;

  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  await navigation.navigate("#1").finished;

  let back_destination;
  navigation.addEventListener("navigate", t.step_func(e => {
    back_destination = e.destination;
    assert_equals(back_destination.index, start_index);
  }), { once: true });
  await navigation.back().finished;

  // Disposing the destination entry of back_destination should update
  // back_destination.index, even though back_destination's navigation has
  // completed.
  await navigation.navigate("#clobber_back", { history: "replace" }).finished;
  assert_equals(back_destination.index, -1);

  navigation.addEventListener("navigate", t.step_func(e => {
    assert_equals(e.destination.index, start_index + 1);

    // Dispose the destination entry and destination.index should update.
    navigation.navigate("#clobber_forward");
    assert_equals(e.destination.index, -1);
  }), { once: true });
  await assertBothRejectDOM(t, navigation.forward(), "AbortError");
}, "navigate event destination.index should be dynamic");
</script>
