LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-destination-dynamic-index-expected.txt


FAIL navigate event destination.index should be dynamic assert_equals: expected 1 but got -1
